Перероблено операції переводу чисел в 16 систему числення, додано валідацію в проект Nebesnyi_task.exe

+++ lab08/prj/ModulesNebesnyi.h
@@ -1,6 +1,6 @@
 #ifndef MODULESNEBESNYI_H_INCLUDED
 #define MODULESNEBESNYI_H_INCLUDED
 
-float s_calculation(float x, float y, float z);
+float s_calculation(int x, int y, int z);
 
 #endif // MODULESNEBESNYI_H_INCLUDED
